Remove auto-entropy for SAC Discrete
Removed the automatic entropy tuning implementation from SAC with
discrete actions. We do not use auto-entropy in our experiments, and the
implementation was mistakenly copied from the `SAC.py` file. In reality,
our paper was interested in gauging sensitivity to entropy, which is
difficult when using auto-entropy tuning.

Auto-entropy in the discrete-action setting actually introduces two new
hyperparameters to tune:

1. Entropy scale stepsize
2. Entropy target

whereas with a set entropy scale, we just have a single hyperparameter
to tune.

For continuous action environments, we can use the SAC author's
suggestion for determining the entropy scale target.